perf: reduce bundle size and runtime overhead
