+++ docs/architecture.md
@@ -15,7 +15,7 @@ Vite（`vite.config.ts` 的 vite-plugin-electron/simple）把 `electron/main.ts`
 | 模块                  | 职责                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         |
 | --------------------- | ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | -------- | ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `manager.ts`          | `AgentManager`：编排 sidecar。`promptStreaming()`（`session.promptAsync` 非阻塞，agent=`lumo`、model 传 `{providerID:"oomol", modelID:"oopilot"}` 组合对象（下文 `oomol/oopilot` 即其简写），每轮把 `buildAuthorizedSystem()` 注入 `body.system`——R4 已授权 Link provider 存在性提示，实测为**追加**到 agent.prompt 之后而非覆盖；默认不列具体 provider 名，避免可用性上下文变成工具诱导）；`sendMessage()`（阻塞，headless 用）；`subscribe()`（OpenCode 全局 SSE 事件循环）；session CRUD；`listAuthorizedServices()` 直查 `${connectorBaseUrl}/v1/apps`。                                                                                                                                                                                                                                                                 |
-| `sidecar.ts`          | `OpencodeSidecar`：spawn `opencode serve --hostname=127.0.0.1 --port=0`，解析 stdout 中 "listening on URL" 拿地址。**禁止加 `--pure`**（会跳过自定义插件，`.opencode/tools` 三个连接器工具会静默失效）。配置经 `OPENCODE_CONFIG_CONTENT` 环境变量内联注入（apiKey 只入内存 env，不落盘）；`OPENCODE_CONFIG_DIR` / `XDG_CONFIG_HOME` / `XDG_DATA_HOME` 指向 `userData/agent/isolation`（隔离全局 `~/.config/opencode`；目录须 `mkdirSync` 预建，缺失会 500）；随机 `OPENCODE_SERVER_PASSWORD` Basic Auth；`dispose()` SIGTERM。                                                                                                                                                                                                                                                                                               |
+| `sidecar.ts`          | `OpencodeSidecar`：spawn `opencode serve --hostname=127.0.0.1 --port=0`，解析 stdout 中 "listening on URL" 拿地址。**禁止加 `--pure`**（会跳过自定义插件，`.opencode/tools` 三个连接器工具会静默失效）。配置经 `OPENCODE_CONFIG_CONTENT` 环境变量内联注入（apiKey 只入内存 env，不落盘）；`OPENCODE_CONFIG_DIR` / `XDG_CONFIG_HOME` / `XDG_DATA_HOME` 指向 `userData/agent/isolation`（隔离全局 `~/.config/opencode`；目录须在启动前异步预建，缺失会 500）；随机 `OPENCODE_SERVER_PASSWORD` Basic Auth；`dispose()` SIGTERM。                                                                                                                                                                                                                                                                                                |
 | `config.ts`           | `buildOpencodeConfig()`：provider `oomol`（npm `@ai-sdk/openai-compatible`，baseURL=`llmBaseUrl`）、model `oopilot`（注意：网关 `/v1/models` **不会列出** `oopilot`，它是网关侧别名、由 chat/completions 路由到真实模型——勿据 models 列表"纠正"模型名）、agent `lumo`（prompt 整段替换为 `LUMO_SYSTEM_PROMPT`）。`LUMO_PERMISSION` 全 allow（edit/bash/webfetch/external_directory），同时下发到 agent 级与根级；不下发 tools 禁用表（内置工具全部启用）。                                                                                                                                                                                                                                                                                                                                                                   |
 | `system-prompt.ts`    | `LUMO_SYSTEM_PROMPT`（英文）：Worker / task-first 定位——先判断用户要完成的工作结果，再在直接回答、Local tools、Link tools 之间选最短可靠路径；需要本机上下文时用 Local tools（bash、文件、脚本、具体 URL；cwd 是私有 scratch，访问真实文件用绝对路径或 `~`）；只有任务确实需要连接账号/SaaS 数据或动作时才进 Link tools（search→inspect→call 流程）。保留 inspect-before-call、authorization_required 阻断、最小访问/载荷、副作用确认等契约。                                                                                                                                                                                                                                                                                                                                                                                |
 | `tool-sources.ts`     | 3 个自定义工具的 **String.raw 内嵌源码**（`search_actions.ts` / `inspect_action.ts` / `call_action.ts`），导出 `AGENT_TOOL_FILES`。运行于 OpenCode 的 Bun，不参与本项目 tsc/oxlint；经 `execFile` 调 oo（路径取 `process.env.LUMO_OO_BIN`，回退 `"oo"`——注意打包/Finder 启动的 GUI 进程**不继承 shell PATH**（PATH 为空），二进制一律走绝对路径（`LUMO_OO_BIN` env / `resolveBundledBin`），`"oo"` 字符串回退仅是 dev 兜底，生产不可用）。`call_action` 把 stderr 中 `errorCode: <code>` 命中 AUTH_BLOCKING 集合的错误翻译为 `{status:"authorization_required", authUrl: <console>/app-connections?provider=...}`（authUrl base 取 `LUMO_CONSOLE_URL` env，源码内仍留 `https://console.oomol.com` 字面量 fallback——这是公开主域、非 `oomol.dev`，不触犯 R2 的发布 grep 不变式，属有意保留的最后兜底，运行时总被 env 覆盖）。 |
