+++ electron/chat/node.ts
@@ -5,15 +5,11 @@ import type {
   AgentRuntimeStatus,
   AttachmentPreviewRequest,
   AttachmentPreviewResult,
-  BillingLogItem,
   BillingOverviewRequest,
   BillingOverviewResult,
   BillingSummaryResult,
-  BillingSpendStats,
-  CreditItem,
   ChatMessage,
   ChatService,
-  CreditUsages,
   CreditBalanceResult,
   ChatContextMention,
   LocalArtifactPreviewRequest,
@@ -32,8 +28,6 @@ import type {
   OpenTopUpCheckoutRequest,
   ResolveLocalArtifactsRequest,
   ResolveLocalArtifactsResult,
-  SubscriptionSchedule,
-  SubscriptionStatus,
   SendMessageRequest,
   TranscribeVoiceRequest,
   TranscribeVoiceResult,
@@ -47,7 +41,7 @@ import { open, readdir, readFile, realpath, stat } from "node:fs/promises"
 import os from "node:os"
 import path from "node:path"
 import { translateOpencodeEvent } from "../agent/event-translator.ts"
-import { consoleBaseUrl, consoleServerBaseUrl, insightBaseUrl, voiceAsrBaseUrl } from "../domain.ts"
+import { voiceAsrBaseUrl } from "../domain.ts"
 import { ServiceEvent } from "../service-events.ts"
 import { applyArtifactRoots, recordArtifactRoot } from "./artifact-roots.ts"
 import {
@@ -57,22 +51,18 @@ import {
   mimeFromPath,
   normalizeLocalPathCandidate,
 } from "./artifacts.ts"
+import { BillingClient } from "./billing.ts"
 import { ChatService as ChatServiceName } from "./common.ts"
 import { normalizeChatError } from "./error.ts"
 import { applyStoppedGenerations, recordStoppedGeneration } from "./stopped-generations.ts"
+
+export { billingLogRanges, readBillingLogs } from "./billing.ts"
 
 const attachmentPreviewMaxBytes = 16 * 1024 * 1024
 const artifactTextPreviewMaxBytes = 512 * 1024
 const artifactManifestFileName = ".lumo-artifact.json"
 const userStopAbortWindowMs = 30_000
 const defaultMaxDirectoryItems = 80
-const billingPath = "/billing"
-const dayMs = 24 * 60 * 60 * 1000
-const billingRequestTimeoutMs = 12_000
-const billingLogsMaxRangeDays = 30
-const billingLogsMaxPagesPerRange = 100
-const billingSummaryCacheMs = 30_000
-const billingOverviewCacheMs = 60_000
 
 function errorMessage(error: unknown): string {
   if (error instanceof Error) {
@@ -147,213 +137,9 @@ export function buildContextMentionsSystem(mentions: ChatContextMention[] | unde
   return lines.join("\n")
 }
 
-function formatCredits(value: unknown): string | null {
-  const amount = typeof value === "string" || typeof value === "number" ? Number(value) : Number.NaN
-  if (!Number.isFinite(amount)) {
-    return null
-  }
-  return `$${new Intl.NumberFormat(undefined, { maximumFractionDigits: amount >= 100 ? 0 : 2 }).format(amount)}`
-}
-
-function sumCreditValues(values: unknown[]): number {
-  return values.reduce<number>((sum, value) => {
-    const amount = typeof value === "string" || typeof value === "number" ? Number(value) : Number.NaN
-    return Number.isFinite(amount) ? sum + amount : sum
-  }, 0)
-}
-
-function isGeneralCreditItem(item: unknown): boolean {
-  if (!item || typeof item !== "object") {
-    return true
-  }
-  const scope = "serviceScope" in item && typeof item.serviceScope === "string" ? item.serviceScope : ""
-  const normalized = scope
-    .trim()
-    .toLowerCase()
-    .replace(/[\s_-]+/g, "")
-  if (!normalized) {
-    return true
-  }
-  if (new Set(["all", "common", "default", "general", "global", "universal", "通用"]).has(normalized)) {
-    return true
-  }
-  return !/auth|authorization|authorisation|link|cloud|授权|链接|云任务/.test(normalized)
-}
-
-function filterGeneralCreditUsages(usages: CreditUsages): CreditUsages {
-  const items = usages.items.filter(isGeneralCreditItem)
-  return {
-    ...usages,
-    items,
-    total: {
-      originalCredit: String(sumCreditValues(items.map((item) => item.originalCredit))),
-      currentCredit: String(sumCreditValues(items.map((item) => item.currentCredit))),
-    },
-  }
-}
-
-function readCreditBalance(payload: unknown): CreditBalanceResult {
-  const record = payload && typeof payload === "object" ? (payload as Record<string, unknown>) : {}
-  const items = Array.isArray(record["items"]) ? record["items"].filter(isGeneralCreditItem) : []
-  const total =
-    record["total"] && typeof record["total"] === "object" ? (record["total"] as Record<string, unknown>) : {}
-  const rawCurrent =
-    items.length > 0
-      ? sumCreditValues(
-          items.map((item) =>
-            item && typeof item === "object" ? (item as Record<string, unknown>)["currentCredit"] : undefined,
-          ),
-        )
-      : total["currentCredit"]
-  const amount = typeof rawCurrent === "number" ? rawCurrent : Number(rawCurrent)
-  return {
-    balance: formatCredits(rawCurrent),
-    hasCredits: Number.isFinite(amount) && amount > 0,
-  }
-}
-
-function readCreditUsages(payload: unknown): CreditUsages {
-  const record = payload && typeof payload === "object" ? (payload as Record<string, unknown>) : {}
-  const total =
-    record["total"] && typeof record["total"] === "object" ? (record["total"] as Record<string, unknown>) : {}
-  return {
-    items: Array.isArray(record["items"])
-      ? (record["items"].filter((item): item is CreditItem =>
-          Boolean(item && typeof item === "object"),
-        ) as CreditItem[])
-      : [],
-    ...(typeof record["nextToken"] === "string" ? { nextToken: record["nextToken"] } : {}),
-    total: {
-      originalCredit: String(total["originalCredit"] ?? "0"),
-      currentCredit: String(total["currentCredit"] ?? "0"),
-    },
-    deficit: String(record["deficit"] ?? "0"),
-  }
-}
-
-export function readBillingLogs(payload: unknown): BillingLogItem[] {
-  const source = unwrapApiData<unknown>(payload)
-  if (Array.isArray(source)) {
-    return source.filter(isBillingLogItem)
-  }
-  if (!source || typeof source !== "object") {
-    return []
-  }
-  const record = source as Record<string, unknown>
-  const items = [record["items"], record["logs"], record["records"]].find(Array.isArray)
-  return Array.isArray(items) ? items.filter(isBillingLogItem) : []
-}
-
-function isBillingLogItem(item: unknown): item is BillingLogItem {
-  return Boolean(item && typeof item === "object")
-}
-
-function ensureHttpUrl(rawUrl: string): string {
-  const url = new URL(rawUrl)
-  if (url.protocol !== "http:" && url.protocol !== "https:") {
-    throw new Error("Only http and https URLs can be opened.")
-  }
-  return url.toString()
-}
-
-function billingUrl(target: OpenBillingPageRequest["target"]): string {
-  const url = new URL(billingPath, consoleBaseUrl)
-  if (target === "usage") {
-    url.searchParams.set("tab", "usage")
-  }
-  return url.toString()
-}
-
-function authRequest(token: string): RequestInit {
-  return {
-    credentials: "include",
-    headers: {
-      Accept: "application/json",
-      Authorization: `Bearer ${token}`,
-      Cookie: `oomol-token=${token}`,
-    },
-  }
-}
-
-function checkoutReturnUrl(): string {
-  const target = new URL(consoleBaseUrl)
-  if (target.hostname.startsWith("console.")) {
-    target.hostname = `chat.${target.hostname.slice("console.".length)}`
-  }
-  target.pathname = billingPath
-  target.search = ""
-  target.hash = ""
-  return target.toString()
-}
-
-function statsRange(days: number): { endTime: number; startTime: number } {
-  const normalizedDays = Number.isFinite(days) && days > 0 ? Math.floor(days) : 30
-  const endTime = Date.now()
-  return { endTime, startTime: endTime - normalizedDays * dayMs }
-}
-
-export interface BillingLogRange {
-  endTime: number
-  startTime: number
-}
-
 interface ChatServiceDeps {
   artifactRootStore?: ArtifactRootStore
   stoppedGenerationStore?: StoppedGenerationStore
-}
-
-interface BillingCacheEntry<T> {
-  accountKey: string
-  data: T
-  fetchedAt: number
-}
-
-interface BillingInFlight<T> {
-  accountKey: string
-  promise: Promise<T>
-}
-
-export function billingLogRanges(days: number, endTime = Date.now()): BillingLogRange[] {
-  const normalizedDays = Number.isFinite(days) && days > 0 ? Math.floor(days) : 30
-  const ranges: BillingLogRange[] = []
-  let remainingDays = normalizedDays
-  let rangeEndTime = endTime
-  while (remainingDays > 0) {
-    const rangeDays = Math.min(remainingDays, billingLogsMaxRangeDays)
-    const startTime = rangeEndTime - rangeDays * dayMs
-    ranges.push({ endTime: rangeEndTime, startTime })
-    rangeEndTime = startTime
-    remainingDays -= rangeDays
-  }
-  return ranges
-}
-
-function unwrapConsoleData<T>(payload: unknown): T {
-  if (payload && typeof payload === "object" && "success" in payload && "data" in payload) {
-    const wrapped = payload as { data: T; message?: unknown; success: unknown }
-    if (wrapped.success === false) {
-      throw new Error(typeof wrapped.message === "string" ? wrapped.message : "Request failed.")
-    }
-    return wrapped.data
-  }
-  return payload as T
-}
-
-function unwrapApiData<T>(payload: unknown): T {
-  if (payload && typeof payload === "object" && "data" in payload) {
-    return (payload as { data: T }).data
-  }
-  return payload as T
-}
-
-function logSettledFailure(label: string, result: PromiseSettledResult<unknown>): void {
-  if (result.status === "rejected") {
-    console.warn("[lumo] billing overview request failed", { label, error: errorMessage(result.reason) })
-  }
-}
-
-function createEmptyBillingOverviewResult(): BillingOverviewResult {
-  return { balance: null, spend: null, metering: null, logs: [], subscription: null, schedules: [] }
 }
 
 export function isAbortErrorMessage(message: string): boolean {
@@ -706,9 +492,7 @@ export class ChatServiceImpl extends ConnectionService<ChatService> implements I
 
   private agent: AgentManager | null
   private voiceAuthToken: string | undefined
-  private billingUserId: string | undefined
-  private readonly billingCache = new Map<string, BillingCacheEntry<BillingOverviewResult>>()
-  private readonly billingInFlight = new Map<string, BillingInFlight<BillingOverviewResult>>()
+  private readonly billing = new BillingClient()
   private bridged = false
   private userStoppedSessions = new Map<string, number>()
   private emittedMessageErrors = new Map<string, Set<string>>()
@@ -759,19 +543,15 @@ export class ChatServiceImpl extends ConnectionService<ChatService> implements I
   /** 登录 / 登出时由 main 更新 Studio ASR 需要的 oomol-token。只在主进程内使用，renderer 不可见。 */
   public setVoiceAuthToken(token: string | undefined): void {
     if (this.voiceAuthToken !== token) {
-      this.clearBillingCache()
+      this.billing.setToken(token)
     }
     this.voiceAuthToken = token
   }
 
   /** 登录 / 登出时由 main 更新额度中心所需上下文。凭证只留在主进程内。 */
   public setBillingAccountContext(context: { token?: string; userId?: string }): void {
-    const previousAccountKey = this.billingAccountKey()
     this.voiceAuthToken = context.token
-    this.billingUserId = context.userId
-    if (this.billingAccountKey() !== previousAccountKey) {
-      this.clearBillingCache()
-    }
+    this.billing.setAccountContext(context)
   }
 
   /** agent 就绪后调用：订阅 OpenCode SSE，转译为 ServerEvents 广播给渲染层。 */
@@ -1170,302 +950,31 @@ export class ChatServiceImpl extends ConnectionService<ChatService> implements I
   }
 
   public async openBillingPage(req: OpenBillingPageRequest): Promise<void> {
-    await shell.openExternal(ensureHttpUrl(billingUrl(req.target)))
+    await shell.openExternal(this.billing.billingPageUrl(req))
   }
 
   public async openTopUpCheckout(req: OpenTopUpCheckoutRequest): Promise<void> {
-    if (!this.voiceAuthToken) {
-      await this.openBillingPage({ target: "recharge" })
-      return
-    }
-    const url = new URL("/api/user/web_top_up_url", consoleServerBaseUrl)
-    url.searchParams.set("price", req.price)
-    url.searchParams.set("redirect", checkoutReturnUrl())
-    const checkoutUrl = unwrapConsoleData<string>(await this.fetchConsoleJson(url))
-    if (!checkoutUrl) {
-      throw new Error("Top-up URL response is invalid.")
-    }
-    await shell.openExternal(ensureHttpUrl(checkoutUrl))
+    await shell.openExternal(await this.billing.topUpCheckoutUrl(req))
   }
 
   public async openSubscriptionCheckout(req: OpenSubscriptionCheckoutRequest): Promise<void> {
-    if (!this.voiceAuthToken) {
-      await this.openBillingPage({ target: "recharge" })
-      return
-    }
-    const url = new URL("/api/user/subscriptions/page", consoleServerBaseUrl)
-    url.searchParams.set("payment_type", "subscription")
-    url.searchParams.set("redirect", checkoutReturnUrl())
-    url.searchParams.set("source_page", checkoutReturnUrl())
-    url.searchParams.set("client_platform", "chat-web")
-    url.searchParams.set("plan", req.plan)
-    if (this.billingUserId) {
-      url.searchParams.set("user_id", this.billingUserId)
-    }
-    await shell.openExternal(ensureHttpUrl(url.toString()))
+    await shell.openExternal(await this.billing.subscriptionCheckoutUrl(req))
   }
 
   public async openSubscriptionPortal(): Promise<void> {
-    if (!this.voiceAuthToken) {
-      await this.openBillingPage({ target: "recharge" })
-      return
-    }
-    const url = new URL("/api/stripe/portal", consoleServerBaseUrl)
-    url.searchParams.set("product", "ai")
-    const portalUrl = unwrapConsoleData<string>(await this.fetchConsoleJson(url))
-    if (!portalUrl) {
-      throw new Error("Subscription portal URL response is invalid.")
-    }
-    await shell.openExternal(ensureHttpUrl(portalUrl))
+    await shell.openExternal(await this.billing.subscriptionPortalUrl())
   }
 
   public async getBillingSummary(req: BillingOverviewRequest): Promise<BillingSummaryResult> {
-    if (!this.voiceAuthToken) {
-      return createEmptyBillingOverviewResult()
-    }
-    return this.getCachedBillingResult(`summary:${req.days}`, billingSummaryCacheMs, Boolean(req.forceRefresh), () =>
-      this.fetchBillingSummary(req.days),
-    )
+    return this.billing.getBillingSummary(req)
   }
 
   public async getBillingOverview(req: BillingOverviewRequest): Promise<BillingOverviewResult> {
-    if (!this.voiceAuthToken) {
-      return createEmptyBillingOverviewResult()
-    }
-    return this.getCachedBillingResult(`overview:${req.days}`, billingOverviewCacheMs, Boolean(req.forceRefresh), () =>
-      this.fetchBillingOverview(req.days),
-    )
-  }
-
-  private async fetchBillingSummary(days: number): Promise<BillingSummaryResult> {
-    const [balance, spend, metering] = await Promise.allSettled([
-      this.getAllCreditUsages(),
-      this.getCreditSpendStats(days),
-      this.getCreditMeteringStats(days),
-    ])
-    logSettledFailure("balance", balance)
-    logSettledFailure("spend", spend)
-    logSettledFailure("metering", metering)
-    const criticalResults = [balance, spend, metering]
-    const allCriticalFailed = criticalResults.every((result) => result.status === "rejected")
-    if (allCriticalFailed && balance.status === "rejected") {
-      throw balance.reason
-    }
-    return {
-      balance: balance.status === "fulfilled" ? filterGeneralCreditUsages(balance.value) : null,
-      spend: spend.status === "fulfilled" ? spend.value : null,
-      metering: metering.status === "fulfilled" ? metering.value : null,
-      logs: [],
-      subscription: null,
-      schedules: [],
-    }
-  }
-
-  private async fetchBillingOverview(days: number): Promise<BillingOverviewResult> {
-    const [balance, spend, metering, logs, subscription, schedules] = await Promise.allSettled([
-      this.getAllCreditUsages(),
-      this.getCreditSpendStats(days),
-      this.getCreditMeteringStats(days),
-      this.getBillingLogs(days),
-      this.getSubscriptionStatus(),
-      this.getSubscriptionSchedules(),
-    ])
-    logSettledFailure("balance", balance)
-    logSettledFailure("spend", spend)
-    logSettledFailure("metering", metering)
-    logSettledFailure("logs", logs)
-    logSettledFailure("subscription", subscription)
-    logSettledFailure("schedules", schedules)
-    const criticalResults = [balance, spend, metering]
-    const allCriticalFailed = criticalResults.every((result) => result.status === "rejected")
-    if (allCriticalFailed && balance.status === "rejected") {
-      throw balance.reason
-    }
-    return {
-      balance: balance.status === "fulfilled" ? filterGeneralCreditUsages(balance.value) : null,
-      spend: spend.status === "fulfilled" ? spend.value : null,
-      metering: metering.status === "fulfilled" ? metering.value : null,
-      logs: logs.status === "fulfilled" ? logs.value : [],
-      subscription: subscription.status === "fulfilled" ? subscription.value : null,
-      schedules: schedules.status === "fulfilled" ? schedules.value : [],
-    }
-  }
-
-  private async getCachedBillingResult(
-    key: string,
-    ttlMs: number,
-    forceRefresh: boolean,
-    load: () => Promise<BillingOverviewResult>,
-  ): Promise<BillingOverviewResult> {
-    const accountKey = this.billingAccountKey()
-    if (!accountKey) {
-      return createEmptyBillingOverviewResult()
-    }
-
-    const cached = this.billingCache.get(key)
-    const now = Date.now()
-    if (!forceRefresh && cached?.accountKey === accountKey && now - cached.fetchedAt < ttlMs) {
-      return cached.data
-    }
-
-    const inFlight = this.billingInFlight.get(key)
-    if (!forceRefresh && inFlight?.accountKey === accountKey) {
-      return inFlight.promise
-    }
-
-    const request = load()
-      .then((data) => {
-        if (this.billingAccountKey() === accountKey) {
-          this.billingCache.set(key, { accountKey, data, fetchedAt: Date.now() })
-        }
-        return data
-      })
-      .catch((error: unknown) => {
-        if (cached?.accountKey === accountKey) {
-          console.warn("[lumo] using stale billing cache after refresh failed", { key, error: errorMessage(error) })
-          return cached.data
-        }
-        throw error
-      })
-      .finally(() => {
-        if (this.billingInFlight.get(key)?.promise === request) {
-          this.billingInFlight.delete(key)
-        }
-      })
-
-    this.billingInFlight.set(key, { accountKey, promise: request })
-    return request
-  }
-
-  private billingAccountKey(): string | undefined {
-    return this.billingUserId ?? this.voiceAuthToken
-  }
-
-  private clearBillingCache(): void {
-    this.billingCache.clear()
-    this.billingInFlight.clear()
+    return this.billing.getBillingOverview(req)
   }
 
   public async getCreditBalance(): Promise<CreditBalanceResult> {
-    if (!this.voiceAuthToken) {
-      return { balance: null, hasCredits: false }
-    }
-    const response = await fetch(new URL("/v1/balance/available", insightBaseUrl), authRequest(this.voiceAuthToken))
-    const text = await response.text()
-    if (!response.ok) {
-      throw new Error(`Failed to get credit balance: ${response.status}`)
-    }
-    let payload: unknown
-    try {
-      payload = JSON.parse(text)
-    } catch {
-      payload = undefined
-    }
-    return readCreditBalance(payload)
-  }
-
-  private async fetchConsoleJson(url: URL): Promise<unknown> {
-    if (!this.voiceAuthToken) {
-      throw new Error("Sign in is required.")
-    }
-    const response = await fetch(url, {
-      ...authRequest(this.voiceAuthToken),
-      signal: AbortSignal.timeout(billingRequestTimeoutMs),
-    })
-    const text = await response.text()
-    if (!response.ok) {
-      throw new Error(text || `Request failed with status ${response.status}`)
-    }
-    return text ? (JSON.parse(text) as unknown) : undefined
-  }
-
-  private async fetchInsightJson(url: URL): Promise<unknown> {
-    if (!this.voiceAuthToken) {
-      throw new Error("Sign in is required.")
-    }
-    const response = await fetch(url, {
-      ...authRequest(this.voiceAuthToken),
-      signal: AbortSignal.timeout(billingRequestTimeoutMs),
-    })
-    const text = await response.text()
-    if (!response.ok) {
-      throw new Error(text || `Request failed with status ${response.status}`)
-    }
-    return text ? (JSON.parse(text) as unknown) : undefined
-  }
-
-  private async getAllCreditUsages(): Promise<CreditUsages> {
-    const firstPage = await this.getCreditUsages()
-    const items = [...firstPage.items]
-    let nextToken = firstPage.nextToken
-    while (nextToken) {
-      const nextPage = await this.getCreditUsages(nextToken)
-      items.push(...nextPage.items)
-      nextToken = nextPage.nextToken
-    }
-    return { ...firstPage, items, nextToken: undefined }
-  }
-
-  private async getCreditUsages(nextToken?: string): Promise<CreditUsages> {
-    const url = new URL("/v1/balance/available", insightBaseUrl)
-    if (nextToken) {
-      url.searchParams.set("nextToken", nextToken)
-    }
-    return readCreditUsages(unwrapApiData<unknown>(await this.fetchInsightJson(url)))
-  }
-
-  private async getCreditSpendStats(days: number): Promise<BillingSpendStats> {
-    const { endTime, startTime } = statsRange(days)
-    const url = new URL("/v1/stats/billing", insightBaseUrl)
-    url.searchParams.set("granularity", "daily")
-    url.searchParams.set("startTime", String(startTime))
-    url.searchParams.set("endTime", String(endTime))
-    return unwrapApiData<BillingSpendStats>(await this.fetchInsightJson(url))
-  }
-
-  private async getCreditMeteringStats(days: number): Promise<BillingSpendStats> {
-    const { endTime, startTime } = statsRange(days)
-    const url = new URL("/v1/stats/metering", insightBaseUrl)
-    url.searchParams.set("granularity", "daily")
-    url.searchParams.set("startTime", String(startTime))
-    url.searchParams.set("endTime", String(endTime))
-    return unwrapApiData<BillingSpendStats>(await this.fetchInsightJson(url))
-  }
-
-  private async getBillingLogs(days: number): Promise<BillingLogItem[]> {
-    const ranges = billingLogRanges(days)
-    const pages = await Promise.all(ranges.map((range) => this.getAllBillingLogsInRange(range)))
-    return pages.flat().sort((left, right) => Number(right.createdAt) - Number(left.createdAt))
-  }
-
-  private async getAllBillingLogsInRange(range: BillingLogRange): Promise<BillingLogItem[]> {
-    const items: BillingLogItem[] = []
-    for (let page = 1; page <= billingLogsMaxPagesPerRange; page += 1) {
-      const pageItems = await this.getBillingLogsPage(range, page)
-      if (pageItems.length === 0) {
-        break
-      }
-      items.push(...pageItems)
-    }
-    return items
-  }
-
-  private async getBillingLogsPage({ endTime, startTime }: BillingLogRange, page: number): Promise<BillingLogItem[]> {
-    const url = new URL("/v1/logs/billing", insightBaseUrl)
-    url.searchParams.set("from", String(startTime))
-    url.searchParams.set("to", String(endTime))
-    url.searchParams.set("page", String(page))
-    return readBillingLogs(await this.fetchInsightJson(url))
-  }
-
-  private async getSubscriptionStatus(): Promise<SubscriptionStatus> {
-    const url = new URL("/api/user/subscriptions", consoleServerBaseUrl)
-    return unwrapConsoleData<SubscriptionStatus>(await this.fetchConsoleJson(url))
-  }
-
-  private async getSubscriptionSchedules(): Promise<SubscriptionSchedule[]> {
-    const url = new URL("/api/user/subscriptions/schedulers", consoleServerBaseUrl)
-    return unwrapConsoleData<SubscriptionSchedule[]>(await this.fetchConsoleJson(url))
+    return this.billing.getCreditBalance()
   }
 
   public async transcribeVoice(req: TranscribeVoiceRequest): Promise<TranscribeVoiceResult> {
