+++ package.json
@@ -36,11 +36,9 @@
     "@electron/get": "^5.0.0",
     "@iconify-icons/simple-icons": "^1.2.74",
     "@iconify-icons/tabler": "^1.2.95",
-    "@iconify/react": "^6.0.2",
     "@oomol/connection": "^0.2.28",
     "@oomol/connection-electron-adapter": "^0.2.12",
     "@opencode-ai/plugin": "1.17.3",
-    "@streamdown/code": "^1.1.1",
     "@tailwindcss/vite": "^4.3.0",
     "@types/js-yaml": "^4.0.9",
     "@types/node": "^24.12.3",
